[FEATURE] Add marker IS_READ to inbox messages
When showing the inbox the markers for subject, from, etc. can get
wrapped depending on whether a message has been read or not.

But it is not easily possible to simply add a class to the containing
<tr> which denotes if a message has been read. This patch adds a marker
IS_READ to each message and adds a class attribute to each <tr> of the
inbox message list.

Resolves: #67149
Releases: master
Change-Id: I6750b027c16dd69a4636d0d3c603dfaac8c42b16
Reviewed-on: http://review.typo3.org/39749
Reviewed-by: Simon Schaufelberger <simonschaufi@googlemail.com>
Tested-by: Simon Schaufelberger <simonschaufi@googlemail.com>