[BUGFIX] Select sort order in BE user administration broken
It doesn't work to use the Up/Down icons next to the username column.
When the ordering is changed from the default "ASC" the passed ORDER BY
direction will become quoted in the SQL query. This bugfix uses a switch
statement to determine which of the both possible order directions (ASC
or DESC) to use.

Resolves: #67220
Releases: master
Change-Id: I60774ed055b5a2195f3fbd7e79c67a2ae29bd2bb
Reviewed-on: http://review.typo3.org/39857
Reviewed-by: Simon Schaufelberger <simonschaufi@googlemail.com>
Tested-by: Simon Schaufelberger <simonschaufi@googlemail.com>